+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "cscope-code" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.4]
 ### Added
 - Add commands to query the word without modification.
 - Add the configuration to enable or disable 'Show Call Hierarchy'.